port(View): Almost finished porting View from WPF to Avalonia
Since I wanted to switch from Windows to Linux, I decided to migrate Partlyx to a more modern and cross-platform framework so that I could continue developing this application on the new system without any problems. The WPF layer has not been removed yet, but most likely will be, as Avalonia does an excellent job of handling all of its tasks. Almost all parts of the View layer have been adapted to Avalonia. At the moment, the task of completely eliminating errors at the compilation stage has been completed. The remaining issues will be resolved in the future.
- A small update: before migrating the application to Avalonia, classes for working with resource icons and recipes, as well as their settings, were added to the lower layers of the application.