feat(All the layers): Started updating of the hierarchy of parts
It was decided to completely rethink the hierarchy of resources, recipes, and elements. The plan is to remove direct ownership of resources by recipes and instead make recipes independent, and replace ParentResource with a list of outgoing components. This will remove a huge number of restrictions on creating functionality and make it possible to create reversible recipes.
The hierarchy is now as follows:
-- Resources - not included in the hierarchy
-- Recipes - own components (input and output)
-- Components - are instances of resources

Updates:
- All classes in Core, Services (dto), and ViewModels have been changed to match the new hierarchy
- Storage of elements in the database has been changed
- The need to specify parent elements for components in services has been removed
- The IsReversible parameter for recipes and the IsOutput parameter for components have been added
- RecipeExtensions has been added to services to find components with the necessary resources
- A global RecipesVMContainer has been added